+++ src/index.ts
@@ -1,0 +1,2 @@
+export * from './app';
+export * from './ical-generator';